+++ tests/test_feat.py
@@ -4,3 +4,4 @@ or for debug of the feature.
 run test with: make feat
 """
 
+# from typing_extensions import TypeAlias